requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.57.0
django-crispy-forms==2.0
django-filter==23.3
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
idna==3.4
inflection==0.5.1
jsonschema==3.2.0
kombu==5.3.2
oauthlib==3.2.2
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.8
pycparser==2.21
PyJWT==2.8.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
spectacular==0.4.0
sqlparse==0.4.4
tzdata==2023.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.8
